Use trie for improved substitutions - code reorg
